rag-app/rag-s3-local/requirements.txt
langchain
langchain-community
langchain-core
langchain-google-genai
faiss-cpu
sentence-transformers
pypdf
docx2txt
python-dotenv
boto3